Restructured tokens into module folder
